Refactor editing components into their own file
