+++ frontend/src/components/DetailView/Context/DetailContext.tsx
@@ -1,5 +1,5 @@
 import { ReactNode, createContext, useState, JSX } from 'react'
-import { DropdownOption } from '../common/FormComponents'
+import { DropdownOption } from '../common/editingComponents'
 
 export type ModeType = 'read' | 'new' | 'edit'
