+++ frontend/src/components/DetailView/common/FormComponents.tsx
@@ -1,24 +1,6 @@
-import {
-  Card,
-  Typography,
-  Box,
-  Grid,
-  Divider,
-  Modal,
-  Button,
-  TextField,
-  FormControl,
-  InputLabel,
-  MenuItem,
-  Select,
-  SelectChangeEvent,
-  FormControlLabel,
-  Radio,
-  RadioGroup,
-} from '@mui/material'
+import { Card, Typography, Box, Grid, Divider, Modal, Button, TextField } from '@mui/material'
 import { ReactNode, useState, ChangeEvent } from 'react'
 import { useDetailContext } from '../hooks'
-import { RegisterOptions, FieldValues, UseFormRegisterReturn, FieldErrors } from 'react-hook-form'
 
 export const ArrayToTable = ({ array, half }: { array: Array<Array<ReactNode>>; half?: boolean }) => {
   const maxRowLength = Math.max(...array.map(row => row.length))
@@ -147,101 +129,6 @@ export const EditableTextField = <T extends object>({
   )
 
   return <DataValue<T> field={field} EditElement={editingComponent} />
-}
-
-export type DropdownOption = { value: string; display: string }
-
-export const DropdownSelector = <T extends object>({
-  options,
-  name,
-  field,
-}: {
-  options: Array<DropdownOption | string>
-  name: string
-  field: keyof T
-}) => {
-  const { data, setEditData, editData } = useDetailContext<T>()
-  const editingComponent = (
-    <FormControl size="small">
-      <InputLabel id={`${name}-multiselect-label`}>{name}</InputLabel>
-      <Select
-        labelId={`${name}-multiselect-label`}
-        label={name}
-        id={`${name}-multiselect`}
-        value={(editData[field] || '') as string}
-        onChange={(event: SelectChangeEvent) => setEditData({ ...editData, [field]: event.target.value })}
-        sx={{ width: '12em' }}
-        size="small"
-      >
-        {options.map(item => (
-          <MenuItem key={getValue(item)} value={getValue(item)}>
-            {getDisplay(item)}
-          </MenuItem>
-        ))}
-      </Select>
-    </FormControl>
-  )
-
-  const option = options.find(option => getValue(option) === data[field])
-  const displayValue = option ? getDisplay(option) : null
-  return <DataValue<T> field={field} EditElement={editingComponent} displayValue={displayValue} />
-}
-
-export const FormTextField = <T extends string>({
-  register,
-  errors,
-  fieldName,
-  label,
-  required = false,
-}: {
-  register: (name: T, options?: RegisterOptions<FieldValues, T> | undefined) => UseFormRegisterReturn<T>
-  errors: FieldErrors<FieldValues>
-  fieldName: T
-  label: string
-  required: boolean
-}) => (
-  <TextField {...register(fieldName, { required: required })} error={!!errors[fieldName]} {...{ label, required }} />
-)
-
-const getValue = (item: DropdownOption | string) => (typeof item === 'string' ? item : item.value)
-const getDisplay = (item: DropdownOption | string) => (typeof item === 'string' ? item : item.display)
-
-export const RadioSelector = <T extends object>({
-  options,
-  name,
-  field,
-}: {
-  options: Array<DropdownOption | string>
-  name: string
-  field: keyof T
-}) => {
-  const { data, setEditData, editData } = useDetailContext<T>()
-  const editingComponent = (
-    <FormControl>
-      <RadioGroup
-        aria-labelledby={`${name}-radio-selection`}
-        name={name}
-        onChange={(event: ChangeEvent<HTMLInputElement>) =>
-          setEditData({ ...editData, [field]: event?.currentTarget?.value })
-        }
-        value={editData[field]}
-        sx={{ display: 'flex', flexDirection: 'row' }}
-      >
-        {options.map(option => (
-          <FormControlLabel
-            key={getValue(option)}
-            value={getValue(option)}
-            control={<Radio />}
-            label={getDisplay(option)}
-          />
-        ))}
-      </RadioGroup>
-    </FormControl>
-  )
-
-  const option = options.find(option => getValue(option) === data[field])
-  const displayValue = option ? getDisplay(option) : null
-  return <DataValue<T> field={field} EditElement={editingComponent} displayValue={displayValue} />
 }
 
 /*
